+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### **Added**
 - added multi-acc sagemaker-mlops manifest example
+- added bedrock fine tuning module
 
 ### **Changed**
 - fixed model deploy cross-account permissions